refactor: moved dbConnection file from root to db folder
